refactored file based routing and _layout.tsx files
