+++ app/(cabo)/_layout.tsx
@@ -1,0 +1,40 @@
+import { Tabs } from "expo-router";
+import React from "react";
+import {useColorScheme} from "react-native";
+import {TabBarIcon} from "@/components/TabBarIcon";
+import {themeColors} from "@/constants/Colors";
+import {DarkTheme, DefaultTheme, ThemeProvider} from "@react-navigation/native";
+
+export default function CaboTabLayout() {
+    const colorScheme = useColorScheme();
+    const themeText = colorScheme === "light" ? themeColors.light.text : themeColors.dark.text;
+
+    return (
+        <ThemeProvider value={colorScheme === "dark" ? DarkTheme : DefaultTheme}>
+            <Tabs
+                screenOptions={{
+                    tabBarActiveTintColor: themeText.color,
+                    headerShown: false,
+                }}>
+                <Tabs.Screen
+                    name="index"
+                    options={{
+                        title: "Cabo",
+                        tabBarIcon: ({color, focused}) => (
+                            <TabBarIcon name={focused ? "cards" : "cards-outline"} color={color}/>
+                        ),
+                    }}
+                />
+                <Tabs.Screen
+                    name="leaderboard"
+                    options={{
+                        title: "Leaderboard",
+                        tabBarIcon: ({color, focused}) => (
+                            <TabBarIcon name={focused ? "chart-box" : "chart-box-outline"} color={color}/>
+                        ),
+                    }}
+                />
+            </Tabs>
+        </ThemeProvider>
+    );
+}
