+++ app/(cabo)/leaderboard.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/components/leaderboard";
